+++ source/RewardsService/requirements.txt
@@ -1,3 +1,3 @@
 fabric==1.10.2
 motor==1.1.0
-tornado==3.2
+tornado==6.1
