+++ source/RewardsUI/rewards/index.html
@@ -13,7 +13,7 @@
             </tr>
             {% for reward in rewards_data %}
                 <tr>
-                    <td>{{ reward.tier }}<td>
+                    <td>{{ reward.tier }}</td>
                     <td>{{ reward.points }}</td>
                     <td>{{ reward.rewardName }}</td>
                 </tr>
@@ -23,9 +23,9 @@
     </div>
     <div>
         <h2>Add orders</h2>
-        <form>
-            <label>Enter email address: </label><input type="text"/><br>
-            <label>Enter order total: </label><input type="text"/><input type="button" value="Submit Order"/>
+        <form action="{{ add_points_endpoint }}" method="post">
+            <label>Enter email address: </label><input type="text" name="email"/><br>
+            <label>Enter order total: </label><input type="text" name="points"/><input type="submit" value="Submit Order"/>
         </form>
         <hr>
     </div>
@@ -34,7 +34,7 @@
         <form>
             <label>Email address: </label><input type="text"/><input type="button" value="Search"/>
         </form>
-        <table border="1">
+        <table id="data_table" border="1">
             <thead>
             <tr>
                 <th>Email Address</th>
@@ -46,18 +46,30 @@
                 <th>Next Reward Tier Progress</th>
             </tr>
             </thead>
-            <tbody>
-                <tr>
-                    <td>customer01@gmail.com</td>
-                    <td>100</td>
-                    <td>A</td>
-                    <td>5% off purchase</td>
-                    <td>B</td>
-                    <td>10% off purchase</td>
-                    <td>50%</td>
-                </tr>
-            </tbody>
+            
+
+      
         </table>
     </div>
 </body>
+    <script>
+        let populate_data = (data)=>{
+            console.log(data)
+            for (let i = 0; i < data.length; i++) {
+                var table = document.getElementById('data_table');
+                var tr = document.createElement('tr');
+                tr.innerHTML = '<td>'+data[i]['email_address']+'</td>'+
+                    '<td>'+data[i]['reward_points']+'</td>'+
+                    '<td>'+data[i]['next_reward_tier']+'</td>'+
+                    '<td>'+data[i]['next_reward_tier_name']+'</td>'+
+                    '<td>'+data[i]['next_reward_tier']+'</td>'+
+                    '<td>'+data[i]['next_reward_tier_name']+'</td>'+
+                    '<td>'+data[i]['next_reward_tier_progress']+'</td>';
+                    table.appendChild(tr);
+            }
+        }
+        fetch("http://localhost:7050/endpoint_three").then(response =>{return response.json()}).then(data =>{ populate_data(data)})
+        
+        
+    </script>
 </html>
