+++ source/RewardsUI/rewards/views.py
@@ -1,5 +1,6 @@
-import logging
-
+import logging, requests
+from requests.structures import CaseInsensitiveDict
+from rewards.services import get_all_user_rewards
 from django.template.response import TemplateResponse
 from django.views.generic.base import TemplateView
 
@@ -18,6 +19,12 @@ class RewardsView(TemplateView):
 
         rewards_data = self.rewards_service_client.get_rewards()
         context['rewards_data'] = rewards_data
+        context['add_points_endpoint'] = "http://localhost:7050/endpoint_one"
+        try:
+            r = requests.get("http://localhost:7050/endpoint_three")
+        except Exception as e:
+            print("the exception is {}".format(e))
+            
 
         return TemplateResponse(
             request,
